+++ src/world/model.rs
@@ -1,6 +1,5 @@
 
 use std::fmt;
-use gl::types::GLfloat;
 use glm;
 use glm::GenNum;
 use num_traits::One;
@@ -8,19 +7,20 @@ use glm::{ Vector3, Matrix4 };
 
 use graphics::{ create_transformation_matrix };
 use utility::traits::{ Translatable, Rotatable, Scalable };
+use utility::Float;
 
 pub struct Model {
-    position: Vector3<f32>,
-    rotation: Vector3<f32>,
-    scale: Vector3<f32>,
-    matrix: glm::Matrix4<GLfloat>
+    position: Vector3<Float>,
+    rotation: Vector3<Float>,
+    scale: Vector3<Float>,
+    matrix: glm::Matrix4<Float>
 }
 
 impl Model {
     fn update_matrix(&mut self) {
         self.matrix = create_transformation_matrix(self.position, self.rotation, self.scale);
     }
-    pub fn get_matrix(&self) -> Matrix4<GLfloat> {
+    pub fn get_matrix(&self) -> Matrix4<Float> {
         self.matrix.clone()
     }
 }
@@ -28,10 +28,10 @@ impl Model {
 impl Default for Model {
     fn default() -> Self {
         let mut model = Self {
-            position: Vector3::<f32>::from_s(0.),
-            rotation: Vector3::<f32>::from_s(0.),
-            scale: Vector3::<f32>::from_s(1.),
-            matrix: Matrix4::<GLfloat>::one()
+            position: Vector3::<Float>::from_s(0.),
+            rotation: Vector3::<Float>::from_s(0.),
+            scale: Vector3::<Float>::from_s(1.),
+            matrix: Matrix4::<Float>::one()
         };
         model.update_matrix();
         model
@@ -47,18 +47,18 @@ impl fmt::Display for Model {
 }
 
 impl Translatable for Model {
-     fn set_translation(&mut self, new_translation: Vector3<f32>) {
+     fn set_translation(&mut self, new_translation: Vector3<Float>) {
         self.position = new_translation;
         self.update_matrix();
     }
-    fn get_translation(&self) -> Vector3<f32> {
+    fn get_translation(&self) -> Vector3<Float> {
         self.position.clone()
     }
 }
 
 impl Rotatable for Model {
-    fn set_rotation(&mut self, new_rotation: Vector3<f32>) {
-        const DOUBLE_PI: f32 = std::f32::consts::PI * 2.;
+    fn set_rotation(&mut self, new_rotation: Vector3<Float>) {
+        const DOUBLE_PI: Float = std::f32::consts::PI as Float * 2.;
         self.rotation = new_rotation;
         for value in self.rotation.as_array_mut().iter_mut() {
             if *value >= DOUBLE_PI {
@@ -69,14 +69,14 @@ impl Rotatable for Model {
         }
         self.update_matrix();
     }
-    fn get_rotation(&self) -> Vector3<f32> {
+    fn get_rotation(&self) -> Vector3<Float> {
         self.rotation.clone()
     }
 }
 
 impl Scalable for Model {
-     fn set_scale(&mut self, new_scale: Vector3<f32>) {
-        const MIN_SCALE: f32 = 1e-3;
+     fn set_scale(&mut self, new_scale: Vector3<Float>) {
+        const MIN_SCALE: Float = 1e-3;
         self.scale = new_scale;
         for value in self.scale.as_array_mut().iter_mut() {
             if *value < MIN_SCALE {
@@ -85,7 +85,7 @@ impl Scalable for Model {
         }
         self.update_matrix();
     }
-    fn get_scale(&self) -> Vector3<f32> {
+    fn get_scale(&self) -> Vector3<Float> {
         self.scale.clone()
     }
 }
